+++ 3rdparty/tomnomnom/etherpad-lite-client/tools/testcontainer/run-container.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-sudo docker run --detach --publish=9001:9001 $1